zebrad: Create zebrad.toml in acceptance tests from ZebradConfig
Fixes #929
