+++ zebra-test/Cargo.toml
@@ -16,7 +16,6 @@ color-eyre = "0.5"
 tracing = "0.1.19"
 tracing-subscriber = "0.2.11"
 tracing-error = "0.1.2"
-tempdir = "0.3.7"
 spandoc = "0.2.0"
 regex = "1.3.9"
 thiserror = "1.0.20"
