+++ zebra-test/src/prelude.rs
@@ -1,7 +1,5 @@
-pub use crate::command::{tempdir, test_cmd, CommandExt, TestChild};
+pub use crate::command::{test_cmd, CommandExt, TestChild};
 pub use std::process::Stdio;
-
-pub use tempdir::TempDir;
 
 pub use color_eyre;
 pub use color_eyre::eyre;
